COMP: vtkGraphMapper gets/sets floating-point point-size and line-width, to be consistent with the underlying vtkProperty API

+++ VTK/Infovis/vtkGraphMapper.cxx
@@ -201,7 +201,7 @@ void vtkGraphMapper::ColorEdgesOff()
 }
 
 //----------------------------------------------------------------------------
-void vtkGraphMapper::SetVertexPointSize(int size)
+void vtkGraphMapper::SetVertexPointSize(float size)
 {
   this->VertexPointSize = size;
   this->VertexActor->GetProperty()->SetPointSize(this->GetVertexPointSize());
@@ -209,7 +209,7 @@ void vtkGraphMapper::SetVertexPointSize(int size)
 }
   
 //----------------------------------------------------------------------------
-void vtkGraphMapper::SetEdgeLineWidth(int width)
+void vtkGraphMapper::SetEdgeLineWidth(float width)
 {
   this->EdgeLineWidth = width;
   this->EdgeActor->GetProperty()->SetLineWidth(this->GetEdgeLineWidth());
